+++ R/download.R
@@ -249,10 +249,10 @@ downloadFile <- function(archive, targetFile, neededFiles,
                     normPath(fileToDownload),
                     "')\n",
                     "      then rerun this current function call.\n",
-                    if (!is.null(getOption("reproducible.inputPaths"))) {
-                      obj <- dir(getOption("reproducible.inputPaths"), full.names = TRUE, pattern = basename(fileToDownload))
+                    if (!is.null(.getDataPath())) {
+                      obj <- dir(.getDataPath(), full.names = TRUE, pattern = basename(fileToDownload))
                       if (length(obj)) {
-                        paste0(" 2b) The copy of the file in getOption('reproducible.inputPaths')",
+                        paste0(" 2b) The copy of the file in getOption('reproducible.dataPath')",
                                " may have been changed or corrupted -- run:\n",
                                "      file.remove(c('",
                                paste(normPath(obj), collapse = "', '"),